Create ver1.1_NOT_WORKING.c
This is almost identical to version 1, but now: 1. The equations are correct, 2. I have changed the definition of n (number of dry neighbour squares) but this leads to a seg fault when compiling, 3. I tried to change the print frequency in the time loop (t=1 -> t=500) but fmod and % are not working and do not print as often as I want. So, these are things we need to fix. Also, I checked the output with a calculation I performed by hand and they didn't match, so I think is is because of the n definition.